backend/requirements.txt
# API / validation / settings
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
python-dotenv>=1.0,<2.0

# Structured logging / serialization
structlog>=24.0,<26.0
orjson>=3.10,<4.0

# Azure + AI
semantic-kernel>=1.30,<2.0
openai>=1.40,<2.0
azure-identity>=1.17,<2.0
azure-search-documents>=11.6,<12.0
azure-ai-documentintelligence>=1.0,<2.0
azure-storage-blob>=12.22,<13.0

# HTTP / async / retry helpers
httpx>=0.27,<1.0
aiohttp>=3.10,<4.0
tenacity>=8.5,<10.0

# Token budgeting / document rendering / images
tiktoken>=0.8,<1.0
python-docx>=1.1,<2.0
pillow>=10.4,<12.0
lxml>=5.3,<6.0
pyyaml
